feat: add aura_set_max_p99_latency() runtime setter
Allow updating the P99 latency target on an already-created engine,
enabling callers like Caliente to apply per-stream hints after context
creation. The setter iterates all rings under lock and updates the
adaptive controller's max_p99_ms field, taking effect on the next
AIMD tick. Passing 0 reverts to auto-detect.

Also fix pre-existing i32/u32 type mismatch in Rust RequestHandle::op_type()
uncovered by bindgen regeneration.

+++ bindings/rust/aura/src/lib.rs
@@ -1553,7 +1553,7 @@ mod tests {
 
         assert_eq!(
             op,
-            aura_sys::aura_op_type_t_AURA_OP_READ as i32,
+            aura_sys::aura_op_type_t_AURA_OP_READ,
             "op_type should be AURA_OP_READ"
         );
     }
